将danmaku_shift字段类型改为double

+++ src/main/java/com/devteam/mikufunbackend/entity/DandanPlayMatchV0.java
@@ -14,5 +14,5 @@ public class DandanPlayMatchV0 {
     String episodeTitle;
     String type;
     String typeDescription;
-    int shift;
+    double shift;
 }
